# app/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PtLinker API",
        "description": "PT站点爬虫管理API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/statistics": {
            "get": {
                "tags": [
                    "statistics",
                    "statistics"
                ],
                "summary": "获取统计数据",
                "description": "获取站点数据统计信息，支持多种统计指标和聚合方式。\n\n### 参数说明\n- **site_id**: 站点ID，可选。如果不指定，则统计所有站点数据\n- **start_date**: 统计开始日期，格式：YYYY-MM-DD。如果不指定，默认为7天前\n- **end_date**: 统计结束日期，格式：YYYY-MM-DD。如果不指定，默认为今天\n- **metrics**: 要统计的指标列表，可选值：\n    - `daily_results`: 每日最新数据（上传量、下载量、魔力值等）\n    - `daily_increments`: 每日增量数据（上传增量、魔力值增量等）\n    - `checkins`: 签到数据（签到状态、时间等）\n- **include_fields**: 要包含的字段列表，可选。例如：[\"upload\", \"download\", \"bonus\"]\n- **exclude_fields**: 要排除的字段列表，可选。例如：[\"seeding_size\", \"seeding_count\"]\n- **group_by**: 分组维度列表，可选。例如：[\"site_id\", \"date\"]\n- **time_unit**: 时间聚合单位，可选值：\n    - `day`: 按天聚合（默认）\n    - `week`: 按周聚合\n    - `month`: 按月聚合\n- **calculation**: 聚合计算方式，可选值：\n    - `last`: 取最后一条记录（默认）\n    - `max`: 取最大值\n    - `min`: 取最小值\n    - `avg`: 取平均值\n    - `sum`: 取总和\n\n### 返回数据说明\n返回一个 StatisticsResponse 对象，包含：\n- **time_range**: 实际统计的时间范围\n    - start: 开始日期\n    - end: 结束日期\n- **metrics**: 各项统计指标的数据\n    - daily_results: 每日最新数据列表\n    - daily_increments: 每日增量数据列表\n    - checkins: 签到数据列表\n- **summary**: 汇总数据\n    - total_sites: 统计的站点总数\n    - total_upload_increment: 总上传量增量\n    - total_bonus_increment: 总魔力值增量\n    - successful_checkins: 成功签到次数\n- **metadata**: 统计元数据\n    - generated_at: 统计生成时间\n    - applied_filters: 应用的过滤条件\n\n### 使用示例\n1. 获取指定站点最近7天的所有统计数据：\n```\nGET /api/v1/statistics?site_id=example_site\n```\n\n2. 获取所有站点指定日期范围的签到统计：\n```\nGET /api/v1/statistics?start_date=2024-01-01&end_date=2024-01-31&metrics=checkins\n```\n\n3. 获取指定站点的每日上传和魔力值增量：\n```\nGET /api/v1/statistics?site_id=example_site&metrics=daily_increments&include_fields=upload_increment,bonus_increment\n```\n\n4. 按月统计所有站点的数据，并计算平均值：\n```\nGET /api/v1/statistics?time_unit=month&calculation=avg\n```\n\n### 错误处理\n- 如果提供的日期格式不正确，将返回 400 错误\n- 如果请求的站点不存在，将返回相应的错误信息\n- 如果查询过程中发生错误，将返回 500 错误",
                "operationId": "get_statistics_api_v1_statistics_get",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Site Id"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "format": "date"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Start Date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "format": "date"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "End Date"
                        }
                    },
                    {
                        "name": "metrics",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/MetricType"
                            },
                            "default": [
                                "daily_results",
                                "daily_increments",
                                "checkins"
                            ],
                            "title": "Metrics"
                        }
                    },
                    {
                        "name": "include_fields",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Include Fields"
                        }
                    },
                    {
                        "name": "exclude_fields",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Exclude Fields"
                        }
                    },
                    {
                        "name": "group_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Group By"
                        }
                    },
                    {
                        "name": "time_unit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/TimeUnit",
                            "default": "day"
                        }
                    },
                    {
                        "name": "calculation",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/CalculationType",
                            "default": "last"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatisticsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/statistics/last-success": {
            "get": {
                "tags": [
                    "statistics",
                    "statistics"
                ],
                "summary": "获取每个站点最后一次成功任务的数据",
                "description": "获取每个站点最后一次成功任务的数据\n\nArgs:\n    site_id: 可选的站点ID，如果提供则只返回该站点的数据\n    db: 数据库会话\n    \nReturns:\n    Dict[str, Dict]: 包含每个站点最后一次成功任务的数据",
                "operationId": "get_last_success_tasks_api_v1_statistics_last_success_get",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Site Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    },
                                    "title": "Response Get Last Success Tasks Api V1 Statistics Last Success Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CalculationType": {
                "type": "string",
                "enum": [
                    "last",
                    "max",
                    "min",
                    "avg",
                    "sum"
                ],
                "title": "CalculationType",
                "description": "计算方式"
            },
...
            "CheckInResult": {
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date",
                        "title": "Date"
                    },
                    "site_id": {
                        "type": "string",
                        "title": "Site Id"
                    },
                    "checkin_status": {
                        "type": "string",
                        "title": "Checkin Status"
                    },
                    "checkin_time": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Checkin Time"
                    },
                    "task_id": {
                        "type": "string",
                        "title": "Task Id"
                    }
                },
                "type": "object",
                "required": [
                    "date",
                    "site_id",
                    "checkin_status",
                    "checkin_time",
                    "task_id"
                ],
                "title": "CheckInResult",
                "description": "签到统计"
            },
...
            "DailyIncrement": {
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date",
                        "title": "Date"
                    },
                    "site_id": {
                        "type": "string",
                        "title": "Site Id"
                    },
                    "upload_increment": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Upload Increment",
                        "default": 0
                    },
                    "bonus_increment": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bonus Increment",
                        "default": 0
                    },
                    "seeding_score_increment": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Seeding Score Increment",
                        "default": 0
                    },
                    "seeding_size_increment": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Seeding Size Increment",
                        "default": 0
                    },
                    "seeding_count_increment": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Seeding Count Increment",
                        "default": 0
                    },
                    "task_id": {
                        "type": "string",
                        "title": "Task Id"
                    },
                    "reference_task_id": {
                        "type": "string",
                        "title": "Reference Task Id"
                    }
                },
                "type": "object",
                "required": [
                    "date",
                    "site_id",
                    "task_id",
                    "reference_task_id"
                ],
                "title": "DailyIncrement",
                "description": "每日数据增量统计"
            },
            "DailyResult": {
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date",
                        "title": "Date"
                    },
                    "site_id": {
                        "type": "string",
                        "title": "Site Id"
                    },
                    "username": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Username"
                    },
                    "upload": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Upload"
                    },
                    "download": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Download"
                    },
                    "ratio": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ratio"
                    },
                    "bonus": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bonus"
                    },
                    "seeding_score": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Seeding Score"
                    },
                    "seeding_size": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Seeding Size"
                    },
                    "seeding_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Seeding Count"
                    },
                    "task_id": {
                        "type": "string",
                        "title": "Task Id"
                    }
                },
                "type": "object",
                "required": [
                    "date",
                    "site_id",
                    "username",
                    "upload",
                    "download",
                    "ratio",
                    "bonus",
                    "seeding_score",
                    "seeding_size",
                    "seeding_count",
                    "task_id"
                ],
                "title": "DailyResult",
                "description": "每日最后结果统计"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MetricType": {
                "type": "string",
                "enum": [
                    "daily_results",
                    "daily_increments",
                    "checkins"
                ],
                "title": "MetricType",
                "description": "统计指标类型"
            },
...
            "StatisticsMetadata": {
                "properties": {
                    "generated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Generated At"
                    },
                    "applied_filters": {
                        "additionalProperties": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "items": {
                                        "type": "string"
                                    },
                                    "type": "array"
                                }
                            ]
                        },
                        "type": "object",
                        "title": "Applied Filters"
                    }
                },
                "type": "object",
                "required": [
                    "generated_at",
                    "applied_filters"
                ],
                "title": "StatisticsMetadata",
                "description": "统计元数据"
            },
            "StatisticsResponse": {
                "properties": {
                    "time_range": {
                        "additionalProperties": {
                            "type": "string",
                            "format": "date"
                        },
                        "type": "object",
                        "title": "Time Range"
                    },
                    "metrics": {
                        "additionalProperties": {
                            "anyOf": [
                                {
                                    "items": {
                                        "$ref": "#/components/schemas/DailyResult"
                                    },
                                    "type": "array"
                                },
                                {
                                    "items": {
                                        "$ref": "#/components/schemas/DailyIncrement"
                                    },
                                    "type": "array"
                                },
                                {
                                    "items": {
                                        "$ref": "#/components/schemas/CheckInResult"
                                    },
                                    "type": "array"
                                }
                            ]
                        },
                        "type": "object",
                        "title": "Metrics"
                    },
                    "summary": {
                        "$ref": "#/components/schemas/StatisticsSummary"
                    },
                    "metadata": {
                        "$ref": "#/components/schemas/StatisticsMetadata"
                    }
                },
                "type": "object",
                "required": [
                    "time_range",
                    "metrics",
                    "summary",
                    "metadata"
                ],
                "title": "StatisticsResponse",
                "description": "统计响应数据"
            },
            "StatisticsSummary": {
                "properties": {
                    "total_sites": {
                        "type": "integer",
                        "title": "Total Sites"
                    },
                    "total_upload_increment": {
                        "type": "number",
                        "title": "Total Upload Increment"
                    },
                    "total_bonus_increment": {
                        "type": "number",
                        "title": "Total Bonus Increment"
                    },
                    "successful_checkins": {
                        "type": "integer",
                        "title": "Successful Checkins"
                    }
                },
                "type": "object",
                "required": [
                    "total_sites",
                    "total_upload_increment",
                    "total_bonus_increment",
                    "successful_checkins"
                ],
                "title": "StatisticsSummary",
                "description": "统计汇总数据"
            },
...
            "TimeUnit": {
                "type": "string",
                "enum": [
                    "day",
                    "week",
                    "month"
                ],
                "title": "TimeUnit",
                "description": "时间聚合单位"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}